refactor(decompose): split packages/libraries/graph-state/scripts/generate-fixtures.ts into folder
